+++ types/modules.d.ts
@@ -8,27 +8,7 @@ declare module '*.tflite' {
   export default assetId;
 }
 
-declare module '*.obj' {
-  const assetId: number;
-  export default assetId;
-}
-
-declare module '*.mtl' {
-  const assetId: number;
-  export default assetId;
-}
-
 declare module '*.onnx' {
-  const assetId: number;
-  export default assetId;
-}
-
-declare module '*.glb' {
-  const assetId: number;
-  export default assetId;
-}
-
-declare module '*.gltf' {
   const assetId: number;
   export default assetId;
 }